% Generated by roxygen2 (4.1.1): do not edit by hand
% Please edit documentation in R/constraint.equation.R
\name{constraint_equation}
\alias{constraint_equation}
\title{Creates a constraint equation based on the input data frame.}
\usage{
constraint_equation(x, weight.var, match.var, replace)
}
\arguments{
\item{x}{data frame containing needed input data}

\item{weight.var}{character name of the column of the input weights}

\item{match.var}{character vector of names of columns of 'data' we wish to
match on}

\item{replace}{logical indicating whether or not observations weighted in the
       original weight.var are allowed positive weight in the output.}
}
\value{
A list with two named components: A and b, representing the
        components of the constraint equation \eqn{Aw = b}
}
\description{
Given an input data frame, extract the matching vectors and the weight vectors out
of the data frame. Then format the extracted data to form a linear constraint \eqn{Aw=b}
that defines a sample space from which we will later sample from.
}
\author{
David Kane \email{<dave.kane@gmail.com>}
}

