\name{hnr_loop}
\alias{hnr_loop}
\title{Loop for hitandrun}
\arguments{
  \item{y_s:}{starting point}

  \item{Z_s:}{null-space matrix}

  \item{n_s:}{number of outputs}

  \item{skiplength_s:}{number to skip}

  \item{discard_s:}{burninlength}
}
\description{
  Runs the hitandrun algorithm in a loop
}

