\docType{package}
\name{kmatching-package}
\alias{kmatching-package}
\title{A package to generate random matching}
\description{
  \tabular{ll}{ Package \tab kmatching\cr Type: \tab
  Package\cr Version: \tab 0.1-1\cr Date: \tab
  2012-07-18\cr License: \tab BSD\cr }
}
\details{
  Using random-walk methods, kmatching generates vectors
  $x$ that satisfy $Ax =b, x>=0$.
}

