\name{dummy}
\alias{dummy}
\title{Creates a dummy matrix for a vector of character/factor variables.}
\usage{
  dummy(vec)
}
\arguments{
  \item{vec}{vector of character or factor variable}
}
\description{
  Creates a dummy matrix for a vector of character/factor
  variables.
}
\examples{
dummy(letters[1:3])
}
\author{
  David Kane \email{<dave.kane@gmail.com>}
}

