\docType{package}
\name{kmatching-package}
\alias{kmatching-package}
\title{A package to generate random matching}
\description{
  \tabular{ll}{ Package \tab kmatching\cr Type: \tab
  Package\cr Version: \tab 0.1\cr Date: \tab 7-18-2012\cr
  License: \tab BSD\cr }
}
\details{
  Using random-walk methods, kmatching generates vectors
  $x$ that satisfy $Ax =b, x>=0$. This can be used to in
  matching statistical studies.
}

